+++ src/config/shard_settings.rs
@@ -6,7 +6,7 @@ define_setting!(ListenAddrSetting, String, "127.0.0.1:23001");
 define_setting!(ExternalAddrSetting, String, "127.0.0.1:23001");
 define_setting!(LoginServerAddrSetting, String, "127.0.0.1:23000");
 define_setting!(LoginServerConnIntervalSetting, u64, 10_u64);
-define_setting!(NumChannelsSetting, usize, 1_usize);
+define_setting!(NumChannelsSetting, u8, 1_u8);
 define_setting!(MaxChannelPopSetting, usize, 100_usize);
 define_setting!(VisibilityRangeSetting, usize, 1_usize);
 define_setting!(AutosaveIntervalSetting, u64, 5_u64);
